Use std::size_t for register id

+++ simulator/mips/mips_register/mips_register.h
@@ -17,13 +17,10 @@
 #include <utility>
 
 class MIPSRegister {
-    enum RegNum : uint8
+    enum RegNum : size_t
     {
-// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
 #define REGISTER(X) MIPS_REG_ ## X
-// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
 #define CP0_REGISTER(X) MIPS_CP0_REG_ ## X
-// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
 #define CP1_REGISTER(X) MIPS_CP1_REG_ ## X
 #include "mips_register.def"
 #undef REGISTER
@@ -44,22 +41,22 @@ public:
     bool is_mips_hi()    const { return value == MIPS_REG_hi; }
     bool is_mips_lo()    const { return value == MIPS_REG_lo; }
 
-    static constexpr MIPSRegister from_cpu_index( uint8 id) noexcept
+    static constexpr MIPSRegister from_cpu_index( size_t id) noexcept
     {
         return MIPSRegister( RegNum{ id});
     }
 
-    static constexpr MIPSRegister from_cp0_index( uint8 id) noexcept
+    static constexpr MIPSRegister from_cp0_index( size_t id) noexcept
     {
-        return MIPSRegister( RegNum{ narrow_cast<uint8>( MIPS_CP0_REG_Context0 + id)});
+        return MIPSRegister( RegNum{ narrow_cast<size_t>( MIPS_CP0_REG_Context0 + id)});
     }
 
     static constexpr MIPSRegister from_cp1_index( uint8 id) noexcept
     {
-        return MIPSRegister( RegNum{ narrow_cast<uint8>( MIPS_CP1_REG_f0 + id)});
+        return MIPSRegister( RegNum{ narrow_cast<size_t>( MIPS_CP1_REG_f0 + id)});
     }
 
-    static MIPSRegister from_gdb_index( uint8 id)
+    static MIPSRegister from_gdb_index( size_t id)
     {
         if (id < 32u)
             return from_cpu_index( id);
