+++ simulator/modules/core/perf_sim.cpp
@@ -134,7 +134,7 @@ void PerfSim<ISA>::dump_statistics() const
 }
 
 template <typename ISA>
-uint64 PerfSim<ISA>::read_gdb_register( uint8 regno) const
+uint64 PerfSim<ISA>::read_gdb_register( size_t regno) const
 {
     if ( regno == Register::get_gdb_pc_index())
         return get_pc();
@@ -143,7 +143,7 @@ uint64 PerfSim<ISA>::read_gdb_register( uint8 regno) const
 }
 
 template <typename ISA>
-void PerfSim<ISA>::write_gdb_register( uint8 regno, uint64 value)
+void PerfSim<ISA>::write_gdb_register( size_t regno, uint64 value)
 {
     if ( regno == Register::get_gdb_pc_index())
         set_pc( value);
